example/arduino/requirements.txt
flask_serial
appdirs
